refactor: merge auth and event trpc routers
