--- apps/ow-api/src/router.ts
@@ -1,6 +1,0 @@
-import { eventRouter } from "./modules/event/event-router.js"
-import { t } from "./trpc.js"
-
-export const appRouter = t.router({
-  event: eventRouter,
-})
