+++ apps/ow-api/src/trpc.ts
@@ -1,11 +1,9 @@
-import { getLogger } from "@dotkomonline/logger"
-import { z } from "zod"
-
 import { initTRPC } from "@trpc/server"
 
 import { Context } from "./context"
+import { eventRouter } from "./modules/event/event-router";
+import { authRouter } from "./modules/auth/auth-router";
 
-const logger = getLogger(import.meta.url)
 // TODO: Superjson
 export const t = initTRPC.context<Context>().create({
   errorFormatter({ shape }) {
@@ -14,18 +12,8 @@ export const t = initTRPC.context<Context>().create({
 })
 
 export const appRouter = t.router({
-  signin: t.procedure
-    .input(
-      z.object({
-        email: z.string().email(),
-        password: z.string(),
-        challenge: z.string(),
-      })
-    )
-    .mutation((req) => {
-      logger.info(req.input)
-      return { msg: "hello world" }
-    }),
+  event: eventRouter,
+  auth: authRouter
 })
 
 export type AppRouter = typeof appRouter
